update tests and docs mostly related to edgy (#402)
- update edgy related docs
- update edgy related tests
- remove broken test
- remove some bad named classes from test collection

+++ tests/databases/edgy/test_base_user.py
@@ -7,7 +7,7 @@ import pytest
 from esmerald.conf import settings
 from esmerald.contrib.auth.edgy.base_user import AbstractUser
 
-database, models = settings.edgy_registry
+models = settings.edgy_registry
 pytestmark = pytest.mark.anyio
 
 
@@ -32,15 +32,16 @@ async def create_test_database():
     try:
         await models.create_all()
         yield
-        await models.drop_all()
+        if not models.database.drop:
+            await models.drop_all()
     except Exception:
         pytest.skip("No database available")
 
 
 @pytest.fixture(autouse=True)
 async def rollback_transactions():
-    with database.force_rollback():
-        async with database:
+    with models.database.force_rollback():
+        async with models:
             yield
