+++ tests/databases/edgy/test_middleware_data.py
@@ -19,7 +19,7 @@ from esmerald.contrib.auth.edgy.middleware import JWTAuthMiddleware
 from esmerald.security.jwt.token import Token
 from esmerald.testclient import create_client
 
-database, models = settings.edgy_registry
+models = settings.edgy_registry
 pytestmark = pytest.mark.anyio
 
 
@@ -46,21 +46,14 @@ class User(AbstractUser):
 jwt_config = JWTConfig(signing_key=settings.secret_key)
 
 
-@pytest.fixture(autouse=True, scope="module")
+@pytest.fixture(autouse=True, scope="function")
 async def create_test_database():
-    try:
-        await models.create_all()
-        yield
-        await models.drop_all()
-    except Exception:
-        pytest.skip("No database available")
-
-
-@pytest.fixture(autouse=True)
-async def rollback_transactions():
-    with database.force_rollback(False):
-        async with database:
+    with models.database.force_rollback(False):
+        async with models:
+            await models.create_all()
             yield
+            if not models.database.drop:
+                await models.drop_all()
 
 
 def generate_user_token(user: User, time=None):
